Disabled ban added

+++ acceptGJFriendRequest20.php
@@ -7,6 +7,8 @@ $gjp = sqlTrim($_POST["gjp"]);
 $targetAccountID = sqlTrim($_POST["targetAccountID"]);
 $requestID = sqlTrim($_POST["requestID"]);
 
+if(disabled($accountID)) exit("-1");
+
 if(checkGJP($gjp, $accountID)) {
 	$q = $db->prepare("DELETE FROM frRequests WHERE requestID = '$requestID'");
 	$q->execute();
