+++ deleteGJAccComment20.php
@@ -6,6 +6,8 @@ $accountID = sqlTrim($_POST["accountID"]);
 $gjp = sqlTrim($_POST["gjp"]);
 $commentID = sqlTrim($_POST["commentID"]);
 
+if(disabled($accountID)) exit("-1");
+
 if(checkGJP($gjp, $accountID)) {
 	$q = $db->prepare("DELETE FROM comments WHERE commentID = '$commentID' AND accountID = '$accountID'");
 	$q->execute();
