+++ deleteGJLevelUser20.php
@@ -6,6 +6,8 @@ $accountID = sqlTrim($_POST["accountID"]);
 $gjp = sqlTrim($_POST["gjp"]);
 $levelID = sqlTrim($_POST["levelID"]);
 
+if(disabled($accountID)) exit("-1");
+
 if(checkGJP($gjp, $accountID)) {
 	$q = $db->prepare("DELETE FROM levels WHERE levelID = '$levelID'");
 	$q->execute();
